backend/requirements.txt
Flask
pandas
numpy
torch
transformers
nltk
seaborn
matplotlib
scikit-learn
python-pptx
scipy
